Update all commands to inherit from abstract command class
Instead of defining commands as their own functions, they are now defined as classes that inherit from an abstract Command class. This is done to allow each command to inherit some base functionality that all commands will use. Namely this improvement makes the usage and help commands to better as it is easier to define relevant messages within the commands themselves rather than in a separate Usage.md file. This improvement will make it much easier for users to discover the functionality of each command without the use of external resources.