Create automated testing via github action
Change project structure to better allow for automated testing by creating separate source and test directories. Include necessary configuration files for installing the project as a module. As part of this move, store version number in the setup.cfg file and not in the readme.

Create github workflow as a means of continuous integration.