Use type hinting to make function purposes clearer
The use of type hinting allows for more apt variable names, and makes it clearer which input types are valid for a function. Overall this makes it easier to understand the program just by reading it.

Because the program is clearer to read, more specific inputs were provided to each command. Defining common list names outside of the list definition avoids the need for the findList function in the main loop.