Differentiate behavior of info and status commands
The use of the info command on NPCs in an encounter is useful in some edge cases, but overall creating a meaningful and situational difference between the two commands allows for making both better.

The most limiting factor of the info command was the desire to not show info only relevant in an encounter when viewing info from the bestiary, namely current hp. Separating the command from the encounter list removes that restriction. Any combat specific requirements can be fulfilled by the status command instead.