Add nickname command
Allows users to display unique names for each NPC in their encounter. If no nickname is specified the NPC's original name is used. Because of this it's more appropriate to print an NPC's nickname in most cases so the unique name is used. The bestiary name (essentially the type of NPC) is still shown in parenthesis in all menus for clarity.