Increment version number for mark/unmark commands
The past series of commits have been focused on allowing users to add custom notes to NPCs in their encounters with the mark, or note, command. Notes can be removed with the unmark command.

Notes take the form of a mark, an asterisk, next to the NPC name. Marks can be accompanied by a custom note viewable in the status menu of the marked NPC.

It was my intention to merge the status and info commands in this series of commits, but I'm waiting on some feedback before going through with that.