Improve remaining command descriptions
This commit adds extra details to the remaining commands that still have undescribed capabilities.

The usage string may be a more concise way to highlight the availability of the all selector, but this solution should be more understandable to the vast majority of users if they decide to look it up.