Search for load command on startup
While this technically is more robust in that it would always find the load command if it exists in the command list it's unlikely that will become necessary. Personally I think this is just far more descriptive and just happens to be more resilient to failure.

Could potentially ensure the load command existed and fail if not found, but that's unlikely to be a real failure state.